Revert: optimizations (Security concerns) and honoring the design of SessionStorage(refresh Tokens)
*note that we only store refresh Token in user model.
if user cookie is stolen, the attacker can technically verify both access and refresh tokens.
But, if refresh token in stored in db, and original owner changes password, db refresh token will be invalidated.
And when attacker's access token expires, he won't be able to regenerate it.

and why not do it with just access token(storing it in db)? becase in normal cases, when access token expires, user will be prompted to log in again, we dont want that. since ref+acc token combo makes sure that user will pretty much never login again unless password is changed.

Signed-off-by: lqsky7 <catinice@outlook.com>